+++ src/main/java/com/wakeUpTogetUp/togetUp/api/avatar/controller/AvatarSpeechController.java
@@ -1,16 +1,10 @@
-package com.wakeUpTogetUp.togetUp.api.home;
+package com.wakeUpTogetUp.togetUp.api.avatar.controller;
 
-import java.time.LocalDateTime;
-
-import com.wakeUpTogetUp.togetUp.api.alarm.service.AlarmQueryService;
-import com.wakeUpTogetUp.togetUp.api.alarm.dto.response.AlarmTimeLineRes;
 import com.wakeUpTogetUp.togetUp.api.auth.AuthUser;
 import com.wakeUpTogetUp.togetUp.api.avatar.application.AvatarSpeechProvider;
 import com.wakeUpTogetUp.togetUp.api.avatar.dto.response.AvatarSpeechResponse;
 import com.wakeUpTogetUp.togetUp.common.Status;
-import com.wakeUpTogetUp.togetUp.common.annotation.LogExecutionTime;
 import com.wakeUpTogetUp.togetUp.common.dto.BaseResponse;
-
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -18,36 +12,21 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@Tag(name = "홈(home)")
+@Tag(name = "아바타 (Avatar)")
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/app/home")
-public class HomeController {
+@RequestMapping("/app/avatars")
+public class AvatarSpeechController {
 
-    private final AlarmQueryService alarmQueryService;
     private final AvatarSpeechProvider avatarSpeechProvider;
 
-    @LogExecutionTime
-    @Operation(summary = "브리핑 보드 - 알람 타임라인")
-    @GetMapping("/brief-board/alarm/timeline")
-    @ApiResponse(
-            responseCode = "200",
-            description = "요청에 성공하였습니다.",
-            content = @Content(schema = @Schema(implementation = AlarmTimeLineRes.class)))
-    BaseResponse<AlarmTimeLineRes> getAlarmTimeLineOfUser(
-            @Parameter(hidden = true) @AuthUser Integer userId
-    ) {
-        return new BaseResponse<>(Status.SUCCESS, alarmQueryService.getTimeLine(userId, LocalDateTime.now()));
-    }
-
     @Operation(summary = "아바타 대사 불러오기")
-    @GetMapping("/avatars/{avatarId}/speeches")
+    @GetMapping("/{avatarId}/speeches")
     @ApiResponse(
             responseCode = "200",
             description = "요청에 성공하였습니다.",
